refactor: Add rolesToPermissions
This adds a new function rolesToPermissions which accepts an array of
roles and returns an array of permissions. This function allows us to
extract the logic away from the user class and situate it closer to
where the roles and permissions are defined.
